pyproject.toml
[project]
name = "warpsph"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "joblib>=1.5",
    "pillow>=11.2",
    "plyfile>=1.1",
    "pyglet>=2.1",
    "rtree>=1.4",
    "tqdm>=4.67",
    "trimesh>=4.6",
    "usd-core>=25.5",
    "warp-lang>=1.7",
    "zstandard>=0.23",
]
